refactored animation component
